+++ src/AsyncUDP_WT32_ETH01_Debug.h
@@ -7,13 +7,14 @@
   Built by Khoi Hoang https://github.com/khoih-prog/AsyncUDP_WT32_ETH01
   Licensed under MIT license
   
-  Version: 2.0.2
+  Version: 2.0.3
   
   Version Modified By   Date      Comments
   ------- -----------  ---------- -----------
   2.0.0   K Hoang      10/07/2021 Initial coding for WT32_ETH01. Bump up version to v2.0.0 to sync with AsyncUDP v2.0.0
   2.0.1   K Hoang      12/07/2021 Update to use WebServer_WT32_ETH01 v1.2.0
   2.0.2   K Hoang      09/10/2021 Update `platform.ini` and `library.json`
+  2.0.3   K Hoang      29/11/2021 Auto detect ESP32 core version. Optimize library code. Fix bug in examples
  *****************************************************************************************************************************/
 
 #pragma once